move/rename stuff in wgsl-types
